% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_som.R
\name{aes_som}
\alias{aes_som}
\title{aes_som}
\usage{
aes_som(model_som, class, ts = FALSE, ...)
}
\arguments{
\item{model_som}{object of self-organising maps (SOMs) package}

\item{class}{Class of dataset}

\item{ts}{Boolean type to generate plot of timeseries}

\item{...}{cutree_value division number of clusters}
}
\value{
Model data that is used in plots
}
\description{
aesthetics data of som object
}
