% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treino.R
\name{cria_ggsom}
\alias{cria_ggsom}
\title{cria_ggsom}
\usage{
cria_ggsom(redshift_som, class)
}
\arguments{
\item{redshift_som}{Objeto SOMs}

\item{class}{Classe categórica ou numérica}
}
\description{
Cria um tipo SOMs
}
