% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsom_boxplot.R
\name{gather_data}
\alias{gather_data}
\title{gather_data}
\usage{
gather_data(conjunto)
}
\arguments{
\item{conjunto}{saida do treino da MLP}
}
\value{
Retorna o dado tratado para a visualização
}
\description{
junta o conjunto para a visualização de boxplot no grid
}
