% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsom_boxplot.R
\name{gridsom_boxplot}
\alias{gridsom_boxplot}
\title{gridsom_boxplot}
\usage{
gridsom_boxplot(conjunto)
}
\arguments{
\item{conjunto}{conjunto respectivo a predição da MLP}
}
\value{
Visualização de dados
}
\description{
Visualização de dados da saida da MLP
}
