% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsom_line.R
\name{gridsom_line}
\alias{gridsom_line}
\title{gridsom_line}
\usage{
gridsom_line(redshift_ggsom)
}
\arguments{
\item{redshift_ggsom}{data.frame tratado pela função ggsom}
}
\value{
Visualização dos grids da rede SOMs
}
\description{
Visualização dos neurônios do SOMs utilizando ggplot2
}
\author{
Felipe
}
