% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsom_line.R
\name{junta_conjunto}
\alias{junta_conjunto}
\title{junta_conjunto}
\usage{
junta_conjunto(redshift_ggsom)
}
\arguments{
\item{redshift_ggsom}{data.frame tratado pela função ggsom}
}
\value{
conjunto de dados unido
}
\description{
Necessário para a visualização de dados
}
