% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treino.R
\name{nnet_treino}
\alias{nnet_treino}
\title{nnet_treino}
\usage{
nnet_treino(redshift_ggsom, path, it)
}
\arguments{
\item{path}{Caminho para a escrita do arquivo}

\item{it}{número de iterações}

\item{redshit_ggsom}{data.frame tratado pela função ggsom}
}
\value{
Predição dos Sigmas
}
\description{
Treino da rede neural MLP
}
\author{
Amita
}
