% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treino.R
\name{seleciona_grid}
\alias{seleciona_grid}
\title{seleciona_grid}
\usage{
seleciona_grid(conjunto_dados, posicao_grid)
}
\value{
Conjunto de dados filtrado por X neurônio
}
\description{
Seleciona todos atributos e filtra por um determinado grid
}
