% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treino_mlp.R
\name{treino_mlp}
\alias{treino_mlp}
\title{treino_mlp}
\usage{
treino_mlp(redshifts)
}
\arguments{
\item{redshift}{data.frame do conj redshift}
}
\value{
Predição dos Sigmas
}
\description{
Treino da rede neural MLP
}
\author{
Amita
}
